Done some work on changing the way we send messages back. Also some refactoring.
For sending messages: I am now using response_url across the board to send results back to Slack - this allows the app to send messages anywhere without any restrictions on provate channels/etc. Also there is no need to add the app to any channel as well this way. The only thing that is not supported yet is using commands in Threads in Slack.